marketing/requirements.txt
tweepy==4.14.0
beautifulsoup4==4.12.3
python-dateutil==2.8.2
pytz==2024.1